feat: Combine Question Summary and Response Analysis into single view
Remove the tabbed layout and render metrics, themes, and responses on
one scrollable page.

- Remove QuestionSummary and ResponseAnalysis components
- Remove TabNames enum (no longer needed)
- Remove theme filter dropdown from responses (themes table handles this)
- Replace "Results are filtered" banner with compact inline indicator
- Unify selected highlight style between ThemesTable and MultiChoice
