--- frontend/src/components/dashboard/QuestionSummary/QuestionSummary.test.ts
@@ -1,55 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { render, screen } from "@testing-library/svelte";
-
-import QuestionSummary from "./QuestionSummary.svelte";
-
-describe("QuestionSummary", () => {
-  const demographics = [
-    { id: "1", name: "testCategory", value: "foo", count: 1 },
-  ];
-
-  const testData = {
-    themesLoading: false,
-    questionLoading: false,
-    demographicsLoading: false,
-    showThemes: false,
-    demographics,
-    freeTextResponseCount: 100,
-    themes: [],
-    anyFilterApplied: false,
-    setActiveTab: () => {},
-  };
-
-  it("should render data", () => {
-    render(QuestionSummary, {});
-
-    expect(screen.getByText("Theme analysis")).toBeInTheDocument();
-    expect(
-      screen.getByText(
-        "Analysis of key themes mentioned in responses to this question.",
-      ),
-    ).toBeInTheDocument();
-  });
-
-  it("should render multi choice data if passed", () => {
-    const multiChoice = [
-      { id: "1", text: "yes", response_count: 10 },
-      { id: "2", text: "no", response_count: 20 },
-    ];
-
-    render(QuestionSummary, {
-      ...testData,
-      multiChoice: multiChoice,
-      multiChoiceResponseCount: 50,
-    });
-
-    expect(screen.getByText("Multiple Choice Answers")).toBeInTheDocument();
-
-    multiChoice.forEach((item) => {
-      expect(screen.getByText(item.text)).toBeInTheDocument();
-      expect(screen.getAllByText(item.response_count)).toHaveLength(2);
-      expect(screen.getByText("20%")).toBeInTheDocument();
-      expect(screen.getByText("40%")).toBeInTheDocument();
-    });
-  });
-});
