--- frontend/src/components/dashboard/ResponseAnalysis/ResponseAnalysis.test.ts
@@ -1,17 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { render, screen } from "@testing-library/svelte";
-
-import ResponseAnalysis from "./ResponseAnalysis.svelte";
-
-describe("ResponseAnalysis", () => {
-  it("should render data", () => {
-    render(ResponseAnalysis, {});
-
-    expect(screen.getByText("Response refinement")).toBeInTheDocument();
-    expect(
-      screen.getByText(
-        "In-depth analysis of individual consultation responses.",
-      ),
-    ).toBeInTheDocument();
-  });
-});
